make ClientChannel class definition public
